add tests on ProjectsController #index #show #new #create

+++ app/jobs/expire_project_job.rb
@@ -5,7 +5,7 @@ class ExpireProjectJob < ApplicationJob
     unless project.canceled?
       project.paid? ? project.paid_expired! : project.unpaid_expired!
       project.purge_documents
-      S3Service.delete_object(key: project.zipped_key, bucket: ENV["BUCKET"])
+      S3Service.delete_object(key: project.zipped_key, bucket: ENV["BUCKET"]) if project.zipped_key
       ClientMailer.client_expired(project).deliver_later
       UserMailer.user_expired(project).deliver_later
     end
